code improvements and some new from classes
