--- enums/KeyWords.py
@@ -1,5 +1,0 @@
-from enum import Enum
-
-
-class KeyWords(Enum):
-    PATH_TO_JSON = 'path_to_json'
