--- enums/QuatrainType.py
@@ -1,5 +1,0 @@
-from enum import Enum
-
-
-class QuatrainType(Enum):
-    ABAB = 1
