+++ enums/feature_type.py
@@ -1,0 +1,13 @@
+from enum import Enum
+
+from CustomExceptions import UnsupportedOperationException
+
+
+class FeatureType(Enum):
+    FEATURE = 'Feature'
+
+    def define_by_name(self, string):
+        if string == self.FEATURE:
+            return self.FEATURE
+        else:
+            raise UnsupportedOperationException('Unsupported FeatureType: ' + string)
