+++ enums/key_words.py
@@ -1,0 +1,5 @@
+from enum import Enum
+
+
+class KeyWords(Enum):
+    PATH_TO_JSON = 'path_to_json'
