+++ enums/quatrain_type.py
@@ -1,0 +1,5 @@
+from enum import Enum
+
+
+class QuatrainType(Enum):
+    ABAB = 1
