+++ parser/real_json_sample.json
@@ -1,0 +1,245 @@
+{
+  "type": "FeatureCollection",
+  "features": [
+    {
+      "type": "Feature",
+      "properties": {
+        "MAPBLKLOT": "0001001",
+        "BLKLOT": "0001001",
+        "BLOCK_NUM": "0001",
+        "LOT_NUM": "001",
+        "FROM_ST": "0",
+        "TO_ST": "0",
+        "STREET": "UNKNOWN",
+        "ST_TYPE": null,
+        "ODD_EVEN": "E"
+      },
+      "geometry": {
+        "type": "Polygon",
+        "coordinates": [
+          [
+            [
+              -122.422003528252475,
+              37.808480096967251,
+              0.0
+            ],
+            [
+              -122.422076013325281,
+              37.808835019815085,
+              0.0
+            ],
+            [
+              -122.421102174348633,
+              37.808803534992904,
+              0.0
+            ],
+            [
+              -122.421062569067274,
+              37.808601056818148,
+              0.0
+            ],
+            [
+              -122.422003528252475,
+              37.808480096967251,
+              0.0
+            ]
+          ]
+        ]
+      }
+    },
+    {
+      "type": "Feature",
+      "properties": {
+        "MAPBLKLOT": "0002001",
+        "BLKLOT": "0002001",
+        "BLOCK_NUM": "0002",
+        "LOT_NUM": "001",
+        "FROM_ST": "0",
+        "TO_ST": "0",
+        "STREET": "UNKNOWN",
+        "ST_TYPE": null,
+        "ODD_EVEN": "E"
+      },
+      "geometry": {
+        "type": "Polygon",
+        "coordinates": [
+          [
+            [
+              -122.42082593937107,
+              37.808631474146033,
+              0.0
+            ],
+            [
+              -122.420858049679694,
+              37.808795641369592,
+              0.0
+            ],
+            [
+              -122.419811958704301,
+              37.808761809714007,
+              0.0
+            ],
+            [
+              -122.42082593937107,
+              37.808631474146033,
+              0.0
+            ]
+          ]
+        ]
+      }
+    },
+    {
+      "type": "Feature",
+      "properties": {
+        "MAPBLKLOT": "0004002",
+        "BLKLOT": "0004002",
+        "BLOCK_NUM": "0004",
+        "LOT_NUM": "002",
+        "FROM_ST": "0",
+        "TO_ST": "0",
+        "STREET": "UNKNOWN",
+        "ST_TYPE": null,
+        "ODD_EVEN": "E"
+      },
+      "geometry": {
+        "type": "Polygon",
+        "coordinates": [
+          [
+            [
+              -122.415701204606876,
+              37.808327252671461,
+              0.0
+            ],
+            [
+              -122.415760743593196,
+              37.808630700240904,
+              0.0
+            ],
+            [
+              -122.413787891332404,
+              37.808566801319841,
+              0.0
+            ],
+            [
+              -122.415701204606876,
+              37.808327252671461,
+              0.0
+            ]
+          ]
+        ]
+      }
+    },
+    {
+      "type": "Feature",
+      "properties": {
+        "MAPBLKLOT": "0005001",
+        "BLKLOT": "0005001",
+        "BLOCK_NUM": "0005",
+        "LOT_NUM": "001",
+        "FROM_ST": "206",
+        "TO_ST": "286",
+        "STREET": "JEFFERSON",
+        "ST_TYPE": "ST",
+        "ODD_EVEN": "E"
+      },
+      "geometry": {
+        "type": "Polygon",
+        "coordinates": [
+          [
+            [
+              -122.417346670944355,
+              37.808121127609709,
+              0.0
+            ],
+            [
+              -122.417457443198046,
+              37.808685627252729,
+              0.0
+            ],
+            [
+              -122.416003128921787,
+              37.808638547938997,
+              0.0
+            ],
+            [
+              -122.41593547450509,
+              37.808293744156337,
+              0.0
+            ],
+            [
+              -122.417346670944355,
+              37.808121127609709,
+              0.0
+            ]
+          ]
+        ]
+      }
+    },
+    {
+      "type": "Feature",
+      "properties": {
+        "MAPBLKLOT": "1245061",
+        "BLKLOT": "1245061",
+        "BLOCK_NUM": "1245",
+        "LOT_NUM": "061",
+        "FROM_ST": null,
+        "TO_ST": null,
+        "STREET": null,
+        "ST_TYPE": null,
+        "ODD_EVEN": null
+      },
+      "geometry": null
+    },
+    {
+      "type": "Feature",
+      "properties": {
+        "MAPBLKLOT": "0006001",
+        "BLKLOT": "0006001",
+        "BLOCK_NUM": "0006",
+        "LOT_NUM": "001",
+        "FROM_ST": "350",
+        "TO_ST": "366",
+        "STREET": "JEFFERSON",
+        "ST_TYPE": "ST",
+        "ODD_EVEN": "E"
+      },
+      "geometry": {
+        "type": "Polygon",
+        "coordinates": [
+          [
+            [
+              -122.418974650424474,
+              37.807907564012595,
+              0.0
+            ],
+            [
+              -122.419147512636712,
+              37.808655940108665,
+              0.0
+            ],
+            [
+              -122.418623049836171,
+              37.808723347475208,
+              0.0
+            ],
+            [
+              -122.417699830020709,
+              37.808693471445764,
+              0.0
+            ],
+            [
+              -122.417580761531511,
+              37.808086708494429,
+              0.0
+            ],
+            [
+              -122.418974650424474,
+              37.807907564012595,
+              0.0
+            ]
+          ]
+        ]
+      }
+    }
+  ]
+}
