--- util/Constants.py
@@ -1,1 +1,0 @@
-#store constant here (like path to file)
