+++ util/constants.py
@@ -1,0 +1,1 @@
+#store constant here (like path to file)
